+++ src/main/java/dev/willyelton/crystal_tools/levelable/tool/DiggerLevelableTool.java
@@ -2,19 +2,26 @@ package dev.willyelton.crystal_tools.levelable.tool;
 
 import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.keybinding.KeyBindings;
+import dev.willyelton.crystal_tools.utils.BlockCollectors;
 import dev.willyelton.crystal_tools.utils.LevelUtils;
 import dev.willyelton.crystal_tools.utils.NBTUtils;
+import dev.willyelton.crystal_tools.utils.RayTraceUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.common.Tags;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
 
 /**
  * Class for the shovel, pickaxe, AIOT
@@ -29,14 +36,9 @@ public abstract class DiggerLevelableTool extends LevelableTool {
     @Override
     public boolean mineBlock(@NotNull ItemStack tool, Level level, @NotNull BlockState blockState, @NotNull BlockPos blockPos, @NotNull LivingEntity entity) {
         if (NBTUtils.getFloatOrAddKey(tool, "3x3") > 0 && !NBTUtils.getBoolean(tool, "disable_3x3")) {
-            // TODO: make this actually what I want not this
-            Direction direction = Direction.fromYRot(entity.getYRot());
-            if (entity.getXRot() < -45.0F) {
-                direction = Direction.UP;
-            } else if (entity.getXRot() > 45.0F) {
-                direction = Direction.DOWN;
-            }
-            blockBreakerHelper(tool, level, blockPos, entity, direction);
+            BlockHitResult result = RayTraceUtils.rayTrace(entity);
+            Direction direction = result.getDirection();
+            blockBreakerHelper(tool, level, BlockCollectors.collect3x3(blockPos, direction), entity);
         }
 
         if (NBTUtils.getFloatOrAddKey(tool, "vein_miner") > 0 && KeyBindings.veinMine.isDown() && blockState.is(Tags.Blocks.ORES)) {
@@ -65,41 +67,9 @@ public abstract class DiggerLevelableTool extends LevelableTool {
     }
 
     // just going to hard code it for now
-    private void blockBreakerHelper(ItemStack tool, Level level, BlockPos blockPos, LivingEntity entity, Direction blockFace) {
-        switch (blockFace) {
-            case UP, DOWN -> {
-                // x-z plane break
-                breakBlock(tool, level, blockPos.north(), entity);
-                breakBlock(tool, level, blockPos.south(), entity);
-                breakBlock(tool, level, blockPos.east(), entity);
-                breakBlock(tool, level, blockPos.west(), entity);
-                breakBlock(tool, level, blockPos.north().east(), entity);
-                breakBlock(tool, level, blockPos.north().west(), entity);
-                breakBlock(tool, level, blockPos.south().east(), entity);
-                breakBlock(tool, level, blockPos.south().west(), entity);
-            }
-            case NORTH, SOUTH -> {
-                // x-y or z-y
-                breakBlock(tool, level, blockPos.above(), entity);
-                breakBlock(tool, level, blockPos.below(), entity);
-                breakBlock(tool, level, blockPos.east(), entity);
-                breakBlock(tool, level, blockPos.west(), entity);
-                breakBlock(tool, level, blockPos.east().above(), entity);
-                breakBlock(tool, level, blockPos.west().above(), entity);
-                breakBlock(tool, level, blockPos.east().below(), entity);
-                breakBlock(tool, level, blockPos.west().below(), entity);
-            }
-            case EAST, WEST -> {
-                // x-y or z-y
-                breakBlock(tool, level, blockPos.above(), entity);
-                breakBlock(tool, level, blockPos.below(), entity);
-                breakBlock(tool, level, blockPos.north(), entity);
-                breakBlock(tool, level, blockPos.south(), entity);
-                breakBlock(tool, level, blockPos.north().above(), entity);
-                breakBlock(tool, level, blockPos.south().above(), entity);
-                breakBlock(tool, level, blockPos.north().below(), entity);
-                breakBlock(tool, level, blockPos.south().below(), entity);
-            }
+    private void blockBreakerHelper(ItemStack tool, Level level, List<BlockPos> blockPosCollection, LivingEntity entity) {
+        for (BlockPos pos : blockPosCollection) {
+            breakBlock(tool, level, pos, entity);
         }
     }