+++ changelog.md
@@ -8,6 +8,7 @@
 ### Fixes
 - Fixes Crystal Torches dropping items when placed by a tool
 - Fixes tooltip sticking in the upgrade screen
+- 3x3 Mining is now only based on the face you break not the direction the place is facing
 
 ## 1.4.2 Changes
 ### Changes